+++ src/main/java/com/example/medical_appointment_booking_app/payload/request/Form/ProductForm.java
@@ -7,6 +7,8 @@ import jakarta.validation.constraints.Size;
 import lombok.Data;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductForm {
 
@@ -21,11 +23,11 @@ public class ProductForm {
 
     @NotNull(message = "Quantity must not be null")
     @Min(value = 1, message = "Quantity must be greater than 0")
-    private Long quantity;
+    private int quantity;
 
     @NotNull(message = "Price must not be null")
     @Min(value = 0, message = "Price must be greater than or equal to 0")
-    private Double price;
+    private BigDecimal price;
 
     @NotBlank(message = "Description must not be null or empty")
     private String description;
